refactor: 데이터베이스 초기화 인터페이스 추가
초기화 작업을 위한 `init()` 메서드를 가진 인터페이스 `Initializer`를 추가함.
H2, CSV 데이터베이스 초기화 작업에 다형성을 이용하기 위하여 `Initializer`를 상속하는 `DatabaseInit` 인터페이스를 추가함.
H2, CSV 데이터베이스 초기화 클래스는 이를 구현하도록 하고 빈 객체로 관리함.
