+++ src/main/java/codesquad/application/init/DatabaseInit.java
@@ -1,0 +1,4 @@
+package codesquad.application.init;
+
+public interface DatabaseInit extends Initializer {
+}
