+++ src/main/java/codesquad/application/init/Initializer.java
@@ -1,0 +1,5 @@
+package codesquad.application.init;
+
+public interface Initializer {
+    void init();
+}
